+++ optaplanner-examples/src/main/java/org/optaplanner/examples/common/swingui/latitudelongitude/LatitudeLongitudeTranslator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -84,7 +84,7 @@ public class LatitudeLongitudeTranslator {
     }
 
     public double translateYToLatitude(double y) {
-        return maximumLatitude - ((((double) y) - innerHeightMargin) * latitudeLength / innerHeight);
+        return maximumLatitude - ((y - innerHeightMargin) * latitudeLength / innerHeight);
     }
 
     public int getImageWidth() {
