+++ optaplanner-examples/src/main/java/org/optaplanner/examples/conferencescheduling/persistence/ConferenceSchedulingCfpDevoxxImporter.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -130,7 +130,7 @@ public class ConferenceSchedulingCfpDevoxxImporter {
                 continue;
             }
 
-            TalkType talkType = new TalkType((long) i, talkTypeId);
+            TalkType talkType = new TalkType(i, talkTypeId);
             talkType.setCompatibleRoomSet(new HashSet<>());
             talkType.setCompatibleTimeslotSet(new HashSet<>());
 
@@ -168,7 +168,7 @@ public class ConferenceSchedulingCfpDevoxxImporter {
             int capacity = roomObject.getInt("capacity");
 
             if (!Arrays.asList(IGNORED_ROOM_IDS).contains(id)) {
-                Room room = new Room((long) i);
+                Room room = new Room(i);
                 room.setName(id);
                 room.setCapacity(capacity);
                 room.setTalkTypeSet(getTalkTypeSetForCapacity(capacity));
@@ -209,7 +209,7 @@ public class ConferenceSchedulingCfpDevoxxImporter {
             if (Arrays.asList(IGNORED_SPEAKER_NAMES).contains(speakerName)) {
                 continue;
             }
-            Speaker speaker = new Speaker((long) i);
+            Speaker speaker = new Speaker(i);
             speaker.setName(speakerName);
             speaker.withPreferredRoomTagSet(new HashSet<>())
                     .withPreferredTimeslotTagSet(new HashSet<>())
@@ -299,7 +299,7 @@ public class ConferenceSchedulingCfpDevoxxImporter {
         if (speaker == null) {
             LOGGER.warn("The talk ({}: {}) has a speaker ({}) that doesn't exist in speaker list.", code, title, speakerName);
 
-            speaker = new Speaker((long) solution.getSpeakerList().size());
+            speaker = new Speaker(solution.getSpeakerList().size());
             speaker.setName(speakerName);
             speaker.withPreferredRoomTagSet(new HashSet<>())
                     .withPreferredTimeslotTagSet(new HashSet<>())
@@ -332,7 +332,7 @@ public class ConferenceSchedulingCfpDevoxxImporter {
 
     private void createTalk(String code, String title, String talkTypeId, Set<String> themeTrackSet,
             String languageg, List<Speaker> speakerList, int audienceLevel, Set<String> contentTagSet) {
-        Talk talk = new Talk((long) solution.getTalkList().size());
+        Talk talk = new Talk(solution.getTalkList().size());
         talk.setCode(code);
         talk.setTitle(title);
         if (talkTypeIdToTalkTypeMap.get(talkTypeId) == null) {
